+++ messages/item_exchange_message.go
@@ -6,13 +6,13 @@ import (
 
 type ItemExchangeMessage struct {
 	Message
-	OfferedItem int   `json:"offeredItem"`
-	WantedItem  int   `json:"wantedItem"`
+	OfferedItem int64   `json:"offeredItem"`
+	WantedItem  int64   `json:"wantedItem"`
 	SenderID    int64 `json:"senderId"`
 	Response    bool  `json:"response,omitempty"`
 }
 
-func NewItemExchangeMessage(offered, wanted int, senderID int64) ItemExchangeMessage {
+func NewItemExchangeMessage(offered, wanted int64, senderID int64) ItemExchangeMessage {
 	return ItemExchangeMessage{
 		Message:     NewMessage(ItemExchange),
 		OfferedItem: offered,
@@ -21,7 +21,7 @@ func NewItemExchangeMessage(offered, wanted int, senderID int64) ItemExchangeMes
 	}
 }
 
-func NewItemExchangeMessageWithID(offered, wanted int, id uuid.UUID, senderID int64) ItemExchangeMessage {
+func NewItemExchangeMessageWithID(offered, wanted int64, id uuid.UUID, senderID int64) ItemExchangeMessage {
 	return ItemExchangeMessage{
 		Message:     Message{MessageType: ItemExchange, TransactionID: id},
 		OfferedItem: offered,
